requirements.txt
Brotli==1.0.7
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dash==1.13.3
dash-core-components==1.10.1
dash-html-components==1.0.3
dash-renderer==1.5.0
dash-table==4.8.1
Flask==1.1.2
Flask-Compress==1.5.0
future==0.18.2
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
requests==2.24.0
retrying==1.3.3
six==1.15.0
urllib3==1.25.9
Werkzeug==1.0.1
keras==2.3.1
scikit-learn==0.22.1
pandas==1.0.1
Pillow>=8.1.1
pybase64==1.0.1
numpy==1.18.1
regex==2020.6.8
tensorflow>=2.4.0
